core/requirements.txt
asgiref==3.5.2
# autopep8==1.5.4
black==22.6.0
celery==5.2.7
dj-database-url==0.5.0
Django==3.2.13
django-cacheops==5.0.1
django-celery-beat==2.3.0
django-cleanup==5.0.0
django-cors-headers==3.3.0
django-debug-toolbar==2.2
django-querycount==0.7.0
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.13.1
environs==8.0.0
flower==1.2.0
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-storage==1.30.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
mypy==0.782
mypy-extensions==0.4.3
Pillow==8.2.0
pre-commit==2.20.0
psycopg2-binary==2.8.5
redis==3.5.3
sentry-sdk==0.16.5
whitenoise==5.2.0
